Add files via upload
Question is below
Polynomial Regression:
Generate synthetic data using a function 𝑦 = sin(2𝜋𝑥) + 𝜀, where m=10
(number of training examples) and 𝜀 is a random noise. Apply d-order polynomial model
to the data set where d=0,1,3,9. Plot on the same figure the data and the curves resulting
from your fit for various d (0,1,3,9) values.
Increase the number of training examples to m=100. Repeat the steps done in (I).
Comment briefly on what changed with the increase of m.